feat(multi): Enable command-line option to run without the gui
